src/avatar/tts/xtts_engine.py
# === Файл: src/avatar/tts/xtts_engine.py ===
"""XTTS-v2 TTS Engine implementation (Coqui TTS).

Поддерживает 16+ языков, включая русский, с voice cloning.
Lazy-loading для экономии VRAM (~2GB при загрузке).
"""

from __future__ import annotations

from pathlib import Path
from typing import TYPE_CHECKING

from loguru import logger

from avatar.interfaces.tts import ITTSEngine

if TYPE_CHECKING:
    from collections.abc import AsyncIterator

    from TTS.api import TTS  # pyright: ignore[reportMissingTypeStubs]

    from avatar.schemas.audio_types import AudioSegment


class XTTSEngine(ITTSEngine):
    """XTTS-v2 TTS Engine (Coqui TTS).

    Attributes:
        language: Язык синтеза (по умолчанию "ru").
        speed: Скорость синтеза (1.0 = нормальная).
        speaker_wav: Путь к reference WAV для voice cloning (опционально).
        model: Ленивая загрузка TTS-модели.
    """

    def __init__(
        self,
        language: str = "ru",
        speed: float = 1.0,
        speaker_wav: str | None = None,
    ) -> None:
        """Инициализация XTTS Engine (без загрузки модели).

        Args:
            language: Язык синтеза.
            speed: Скорость синтеза.
            speaker_wav: Путь к WAV для voice cloning.
        """
        self.language = language
        self.speed = speed
        self.speaker_wav = Path(speaker_wav) if speaker_wav else None
        self._model: TTS | None = None
        logger.info(f"XTTSEngine initialized: language={language}, speed={speed}")

    def _load_model(self) -> TTS:
        """Lazy-loading модели XTTS-v2.

        Returns:
            TTS: Загруженная модель Coqui TTS.

        Raises:
            RuntimeError: Если загрузка модели не удалась.
        """
        raise NotImplementedError("TODO: Implement _load_model")

    async def synthesize(
        self,
        text: str,
        language: str = "ru",
        speaker_wav: str | None = None,
    ) -> AudioSegment:
        """Синтез речи из текста (non-streaming).

        Args:
            text: Текст для озвучки.
            language: Язык синтеза.
            speaker_wav: Путь к WAV для voice cloning (перекрывает self.speaker_wav).

        Returns:
            AudioSegment: Аудио-сегмент с метаданными.

        Raises:
            ValueError: Если текст пустой или язык не поддерживается.
            RuntimeError: Если синтез завершился ошибкой.
        """
        raise NotImplementedError("TODO: Implement synthesize")

    async def synthesize_streaming(
        self,
        text: str,
        language: str = "ru",
    ) -> AsyncIterator[AudioSegment]:
        """Синтез речи (streaming).

        Args:
            text: Текст для озвучки.
            language: Язык синтеза.

        Yields:
            AudioSegment: Аудио-чанки.

        Raises:
            ValueError: Если текст пустой.
            RuntimeError: Если синтез завершился ошибкой.
        """
        raise NotImplementedError("TODO: Implement synthesize_streaming")
        yield  # type: ignore[unreachable]

    def get_supported_languages(self) -> list[str]:
        """Список поддерживаемых языков.

        Returns:
            list[str]: Коды языков (например, ["ru", "en", "de", "es", ...]).
        """
        raise NotImplementedError("TODO: Implement get_supported_languages")
